dropbox_backup_to_external_disk_lib

Module compare_mod

Functions

compare_files

compare list: the lists and produce list_for_download, list_for_trash_files

compare_folders

compare folders and write folders to trash into path_list_for_trash_folders the list is already sorted

compare_lists_internal 🔒

compare list: the lists must be already sorted for this to work correctly