dropbox_backup_to_external_disk_lib::dropbox_api_token_with_oauth2_mod

Function decrypt_text_with_metadata

fn decrypt_text_with_metadata(
    encrypted_text_with_metadata: EncryptedTextWithMetadata,
) -> Result<SecretBox<SecretResponseAccessToken>>

decrypt text with metadata

The encrypted file is encoded in base64 just to obfuscate it a little bit.
In json format in plain text there is the “seed”, the private key path and the encrypted secret.
The “seed” will be “signed” with the private key.
Only the “owner” can unlock the private key and sign correctly.
This signature will be used as the true passcode for symmetrical decryption.