dropbox_backup_to_external_disk_lib

Module encrypt_decrypt_mod

Generic functions to encrypt and decrypt secrets using the ssh private key.

Structs

EncryptedTextWithMetadata 🔒

Struct that represents the json data saved in the ‘*.enc’ file.

PathStructInSshFolder

Constants

BLUE

ANSI color

GREEN

ANSI color

RED

ANSI color

RESET

ANSI color

YELLOW

ANSI color

Functions

decode64_from_string_to_32bytes 🔒

Decode base64 from string to 32bytes.

decode64_from_string_to_string 🔒

Decode base64 from string to string.

decrypt_symmetric 🔒

Decrypts plain_encrypted_string with secret_passcode_32bytes.

encode64_from_32bytes_to_string 🔒

Decode base64 from string to 32bytes.

encode64_from_bytes_to_string 🔒

Encode base64 from bytes to string.

encode64_from_string_to_string 🔒

Encode base64 from string to string.

encrypt_symmetric 🔒

Encrypts symmetrically secret_string_to_encrypt with secret_passcode_32bytes.

open_file_b64_get_string 🔒

Get the string from the file that is base64 encoded.

random_seed_32bytes_and_string 🔒

Generate a random seed.

shorten_vec_bytes_to_32bytes 🔒

Shorten the Vec<u8> to [u8;32].

sign_seed_with_private_key_file 🔒

Sign the seed with the private key into 32 bytes secret.

sign_seed_with_ssh_agent 🔒

Sign seed with ssh-agent into 32 bytes secret.

sign_seed_with_ssh_agent_or_private_key_file 🔒

Returns the secret signed seed.