dropbox_backup_to_external_disk_lib::file_txt_mod

Struct FileTxt

pub struct FileTxt {
    file_path: CrossPathBuf,
    file_txt: File,
}

My object to work with text files.

Fields
file_path: CrossPathBuffile_txt: File
Implementations

impl FileTxt

pub fn open_for_read(path: &CrossPathBuf) -> Result<Self>

If file not exist, returns error.

pub fn open_for_read_and_write(path: &CrossPathBuf) -> Result<Self>

If file not exist, it creates it.

pub fn file_path(&self) -> &CrossPathBuf

Returns file path.

pub fn file_name(&self) -> Result<String, LibraryError>

Returns the final component of the Path, if there is one.

pub fn read_to_string(&self) -> Result<String, LibraryError>

This method is similar to fs::read_to_string, but instead of a path it expects a File parameter. \

So is possible to open a File in the bin part of the project and then pass it to the lib part of project.
All input and output should be in the bin part of project and not in the lib.

pub fn write_append_str(&mut self, str: &str) -> Result<()>

Append str to file.

pub fn empty(&mut self) -> Result<()>

Empty the file.