Module dropbox_backup_to_external_disk_lib::local_disk_mod

Module contains all functions for local external disk.

Functions

create_folders

create new empty folders

get_content_hash 🔒

list_local

the logic is in the LIB project, but all UI is in the CLI project they run on different threads and communicate It uses the global APP_STATE for all config data

move_internal 🔒

internal code to move file

move_or_rename_local_files

Files are often moved or renamed After compare, the same file (with different path or name) will be in the list_for_trash_files and in the list_for_download. First for every trash line, we search list_for_download for same size and modified. If found, get the remote_metadata with content_hash and calculate local_content_hash. If they are equal move or rename, else nothing: it will be trashed and downloaded eventually. Remove also the lines in files list_for_trash_files and list_for_download.

move_or_rename_local_files_internal_by_hash 🔒

move_or_rename_local_files_internal_by_name 🔒

read_only_remove

The backup files must not be readonly to allow copying the modified file from the remote. The FileTxt is read+write. It is opened in the bin and not in lib, but it is manipulated only in lib.

trash_files

Move to trash folder the files from list_for_trash_files. Ignore if the file does not exist anymore.

trash_files_internal 🔒

internal

trash_folders

Move to trash folder the folders from list_for_trash_folders.

Type Aliases

ThreadName 🔒