dropbox_backup_to_external_disk_lib

Module remote_dropbox_mod

Module contains all the communication with the remote dropbox storage. It uses inter-thread channels to send text to the UserInterface thread. It uses LibError to return error text to the UI thread.

Structs

DirectoryIterator 🔒

Iterator for Directory on remote Dropbox storage.

Functions

download_from_list

Download files from list. \

download_from_vec 🔒

download_internal 🔒

Download one file with client object dropbox_sdk::default_client::UserAuthDefaultClient.

download_one_file

Download one file is calling internally download_from_vec. \

dropbox_list_folder 🔒

Dropbox function to list folders.

encode_token

This is a short-lived token, so security is not my primary concern. \

get_authorization_token

Read encoded token (from env), decode and return the authorization token.

list_remote

Get remote list in parallel. \

list_remote_folder

List remote folder.

test_connection

Test authentication with dropbox.com. \

Type Aliases

FileList 🔒

FolderList 🔒

FolderListAndFileList 🔒

MasterKey 🔒

ThreadName 🔒

ThreadNum 🔒

TokenEnc 🔒