Module dropbox_backup_to_external_disk_lib::utils_mod

A module with often used functions.

Functions

byte_pos_from_chars

it is used for substring, because string slice are counted in bytes and not chars. if we have multi-byte unicode characters we can get an error if the boundary is not on char boundary.

println_to_ui_thread

println_to_ui_thread sends the string to ui thread and works similarly to println! It panics if there is a bug in the code. This is not a recoverable error.

println_to_ui_thread_with_thread_name

println_to_ui_thread_with_thread_name sends the string to ui thread and works similarly to println! It panics if there is a bug in the code. This is not a recoverable error.

shorten_string

shorten path for screen to avoid word-wrap

sort_list

sort list case insensitive

sort_string_lines

sort string lines case insensitive

Type Aliases

ThreadName 🔒