Module dropbox_backup_to_external_disk_lib::remote_dropbox_mod

Module contains all the communication with the remote dropbox storage. It uses inter-thread channels to send text to the UserInterface thread. It uses LibError to return error text to the UI thread.

Structs

DirectoryIterator 🔒

iterator for Directory on remote Dropbox storage

Functions

download_from_list

download files from list It removes just_downloaded from list_for_download, so this function can be stopped and then called again.

download_from_vec 🔒

download_internal 🔒

download one file with client object dropbox_sdk::default_client::UserAuthDefaultClient

download_one_file

download one file is calling internally download_from_vec This is used just for debugging. For real the user will run download_from_list.

dropbox_list_folder 🔒

dropbox function to list folders

encode_token

This is a short-lived token, so security is not my primary concern. But it is bad practice to store anything as plain text. I will encode it and store it in env var. This is more like an obfuscation tactic to make it harder, but in no way impossible, to find out the secret.

get_authorization_token

read encoded token (from env), decode and return the authorization token

list_remote

get remote list in parallel first get the first level of folders and then request in parallel sub-folders recursively

list_remote_folder

list remote folder

remote_content_hash

get content_hash from remote

test_connection

test authentication with dropbox.com experiment with sending function pointer

Type Aliases

FileList 🔒

FolderList 🔒

FolderListAndFileList 🔒

MasterKey 🔒

ThreadName 🔒

ThreadNum 🔒

TokenEnc 🔒