dropbox_backup_to_external_disk_lib

Trait AppStateMethods

pub trait AppStateMethods: Sync + Send {
    // Required methods
    fn load_keys_from_io(&self) -> Result<(String, String)>;
    fn ref_app_config(&self) -> &AppConfig;
    fn lock_proba(&self) -> MutexGuard<'_, String>;
}

This trait defines what functions must the bin project implement then the lib project can use them. \

All IO must be defined inside the bin project: UI, env, file access.
That way the same lib project can be used from different bin: CLI, TUI, GUI, env, file, network,…
These methods will be available globally.

Required Methods

fn load_keys_from_io(&self) -> Result<(String, String)>

get encrypted authorization token from env var

fn ref_app_config(&self) -> &AppConfig

reference to app_config data

fn lock_proba(&self) -> MutexGuard<'_, String>

get locked Mutex